Bug/scripts (#25)
* Added hyperopt script - and renamed optuna script

* Updates to scripts

Some issues with hyperopt - continuing with optuna for now.

* Changed verbosity [ci skip]

Set verbose output in script

* Bug fix in RNF model

Force long type for inputs - if one of the examples is empty and happens
to be first in the batch, pad_sequence() infers the type as floatTensor
instead of LongTensor - ensure all inputs are of type long to avoid
error in embedding layer

* Renaming of tokenizer

Renamed SimpleTokenizer() to WhiteSpaceTokenizer()

* Updated config file for change in model input

Added optimizer name and args to config. Hydra flexible in regards to
config to add an optimizer arg simply use + in command line.

e.g. +model.optimizer_args.weight_decay=0.00005

Co-authored-by: pbmstrk <pbmstrk@users.noreply.github.com>