Separation of models and LightningModule
Created a clear boundary between models (which now inherit from
nn.Module instead of pl.LightningModule) and Lightning. To do this,
introduced new class TextClassifier which takes an instance of a model -
as such the models can be used without needing to install pytorch
lightning. All the optimizer logic was also moved from the models to the
TextClassifier class.