Separation of lightning and models (#35)
* Added script for TextCNN

* Separation of models and LightningModule

Created a clear boundary between models (which now inherit from
nn.Module instead of pl.LightningModule) and Lightning. To do this,
introduced new class TextClassifier which takes an instance of a model -
as such the models can be used without needing to install pytorch
lightning. All the optimizer logic was also moved from the models to the
TextClassifier class.

* Formatting

* Adjusted scripts for new interface

* Formatting

* Moved config file and added logics for using pre-trained vectors

* Added embedding dropout to textcnn

* minor changes

Co-authored-by: pbmstrk <pbmstrk@users.noreply.github.com>