Major refactor TextClassifier
Instead of passing in name and args and using getattr() to get optimizer
and scheduler inside of class - now pass optimizer and scheduler
explicitly to class. This allows for more complex schedulers that may be
too complex to create using name and args.

Can now also define optimizer and scheduler inside model - in similar
style to a LightningModule